requirements.txt
faiss-cpu # faiss-gpu also fine
pillow
torch
numpy
transformers
platformdirs
accelerate
tqdm